refactor(mcp-server): split proxy runtime
